Made Error standalone
