+++ src/modules/mod.rs
@@ -1,4 +1,5 @@
 mod automata;
+mod error;
 mod parser;
 pub mod regexp;
 mod state;
